requirements.txt
pandas==2.2.2
pandas_ta==0.3.14b0
numpy==1.26.4
python-binance==1.0.19
python-dotenv==1.0.1
setuptools<81